+++ kitty/fonts.c
@@ -346,24 +346,29 @@ adjust_metric(unsigned int *metric, float adj, AdjustmentUnit unit, double dpi) 
     *metric = (a < 0 && -a > (int)*metric) ? 0 : *metric + a;
 }
 
+static unsigned int
+adjust_ypos(unsigned int pos, unsigned int cell_height, int adjustment) {
+    if (adjustment >= 0) adjustment = MIN(adjustment, (int)pos - 1);
+    else adjustment = MAX(adjustment, (int)pos - (int)cell_height + 1);
+    return pos - adjustment;
+}
+
 static void
 calc_cell_metrics(FontGroup *fg) {
     unsigned int cell_height, cell_width, baseline, underline_position, underline_thickness, strikethrough_position, strikethrough_thickness;
     cell_metrics(fg->fonts[fg->medium_font_idx].face, &cell_width, &cell_height, &baseline, &underline_position, &underline_thickness, &strikethrough_position, &strikethrough_thickness);
     if (!cell_width) fatal("Failed to calculate cell width for the specified font");
     unsigned int before_cell_height = cell_height;
-    int cw = cell_width, ch = cell_height;
-    if (OPT(adjust_line_height_px) != 0) ch += OPT(adjust_line_height_px);
-    if (OPT(adjust_line_height_frac) != 0.f) ch = (int)(ch * OPT(adjust_line_height_frac));
-    if (OPT(adjust_column_width_px != 0)) cw += OPT(adjust_column_width_px);
-    if (OPT(adjust_column_width_frac) != 0.f) cw = (int)(cw * OPT(adjust_column_width_frac));
+    unsigned int cw = cell_width, ch = cell_height;
+    adjust_metric(&cw, OPT(cell_width).val, OPT(cell_width).unit, fg->logical_dpi_x);
+    adjust_metric(&ch, OPT(cell_height).val, OPT(cell_height).unit, fg->logical_dpi_y);
 #define MAX_DIM 1000
 #define MIN_WIDTH 2
 #define MIN_HEIGHT 4
     if (cw >= MIN_WIDTH && cw <= MAX_DIM) cell_width = cw;
-    else log_error("Cell width invalid after adjustment, ignoring adjust_column_width");
+    else log_error("Cell width invalid after adjustment, ignoring modify_font cell_width");
     if (ch >= MIN_HEIGHT && ch <= MAX_DIM) cell_height = ch;
-    else log_error("Cell height invalid after adjustment, ignoring adjust_line_height");
+    else log_error("Cell height invalid after adjustment, ignoring modify_font cell_height");
     int line_height_adjustment = cell_height - before_cell_height;
     if (cell_height < MIN_HEIGHT) fatal("Line height too small: %u", cell_height);
     if (cell_height > MAX_DIM) fatal("Line height too large: %u", cell_height);
@@ -373,9 +378,17 @@ calc_cell_metrics(FontGroup *fg) {
 #undef MIN_HEIGHT
 #undef MAX_DIM
 
+    unsigned int baseline_before = baseline;
 #define A(which, dpi) adjust_metric(&which, OPT(which).val, OPT(which).unit, fg->logical_dpi_##dpi);
-    A(underline_thickness, y); A(underline_position, y); A(strikethrough_thickness, y); A(strikethrough_position, y);
+    A(underline_thickness, y); A(underline_position, y); A(strikethrough_thickness, y); A(strikethrough_position, y); A(baseline, y);
 #undef A
+
+    if (baseline_before != baseline) {
+        int adjustment = baseline - baseline_before;
+        baseline = adjust_ypos(baseline_before, cell_height, adjustment);
+        underline_position = adjust_ypos(underline_position, cell_height, adjustment);
+        strikethrough_position = adjust_ypos(underline_position, cell_height, adjustment);
+    }
 
     underline_position = MIN(cell_height - 1, underline_position);
     // ensure there is at least a couple of pixels available to render styled underlines