+++ kitty/options/types.py
@@ -52,9 +52,6 @@ option_names = (  # {{{
  'active_tab_font_style',
  'active_tab_foreground',
  'active_tab_title_template',
- 'adjust_baseline',
- 'adjust_column_width',
- 'adjust_line_height',
  'allow_cloning',
  'allow_hyperlinks',
  'allow_remote_control',
@@ -467,9 +464,6 @@ class Options:
     active_tab_font_style: typing.Tuple[bool, bool] = (True, True)
     active_tab_foreground: Color = Color(0, 0, 0)
     active_tab_title_template: typing.Optional[str] = None
-    adjust_baseline: typing.Union[int, float] = 0
-    adjust_column_width: typing.Union[int, float] = 0
-    adjust_line_height: typing.Union[int, float] = 0
     allow_cloning: choices_for_allow_cloning = 'ask'
     allow_hyperlinks: int = 1
     allow_remote_control: str = 'n'